Move util function in utilities of front-end rather than back-end
