+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewModel.js
@@ -23,7 +23,7 @@ import { jsonFetch } from '../../../utilities/fetch/jsonFetch.js';
 import { RemoteDataSource } from '../../../utilities/fetch/RemoteDataSource.js';
 import { DetectorType } from '../../../domain/enums/DetectorTypes.js';
 import { GaqFilterModel } from '../../../components/Filters/RunsFilter/GaqFilterModel.js';
-import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/stringUtils.js';
+import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/filterOutLegacyDetectorsForNewerPeriods.js';
 
 const ALL_CPASS_PRODUCTIONS_REGEX = /cpass\d+/;
 const DETECTOR_NAMES_NOT_IN_CPASSES = ['EVS'];
