+++ lib/public/views/Runs/RunPerPeriod/RunsPerLhcPeriodOverviewModel.js
@@ -16,7 +16,7 @@ import { jsonFetch } from '../../../utilities/fetch/jsonFetch.js';
 import { DetectorType } from '../../../domain/enums/DetectorTypes.js';
 import { ObservableData } from '../../../utilities/ObservableData.js';
 import { FixedPdpBeamTypeRunsOverviewModel } from '../Overview/FixedPdpBeamTypeRunsOverviewModel.js';
-import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/stringUtils.js';
+import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/filterOutLegacyDetectorsForNewerPeriods.js';
 
 /**
  * Runs Per LHC Period overview model
