+++ lib/public/views/Runs/RunsPerSimulationPass/RunsPerSimulationPassOverviewModel.js
@@ -15,7 +15,7 @@ import { ObservableData } from '../../../utilities/ObservableData.js';
 import { getRemoteData } from '../../../utilities/fetch/getRemoteData.js';
 import { FixedPdpBeamTypeRunsOverviewModel } from '../Overview/FixedPdpBeamTypeRunsOverviewModel.js';
 import { rctDetectorsProvider } from '../../../services/detectors/detectorsProvider.js';
-import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/stringUtils.js';
+import { filterOutLegacyDetectorsForNewerPeriods } from '../../../utilities/filterOutLegacyDetectorsForNewerPeriods.js';
 
 /**
  * Runs Per Simulation Pass overview model
