+++ lib/utilities/index.js
@@ -13,10 +13,8 @@
 
 const deepmerge = require('./deepmerge');
 const isPromise = require('./isPromise');
-const { filterOutLegacyDetectorsForNewerPeriods } = require('./stringUtils');
 
 module.exports = {
     deepmerge,
     isPromise,
-    filterOutLegacyDetectorsForNewerPeriods,
 };
