+++ lib/utilities/stringUtils.js
@@ -72,20 +72,6 @@ const splitStringToStringsTrimmed = (stringCollection, stringSeparator = ',') =>
     .map((string) => string.trim())
     .filter(Boolean);
 
-/**
- * Remove legacy detectors for newer LHC periods.
- *
- * @param {Detector[]} detectors detectors to filter
- * @param {string} [label=''] name of the period or data pass
- * @param {string[]} [legacyDetectorNames=['CPV', 'PHS']] detectors to remove for newer periods after LHC25
- * @return {Detector[]} filtered detectors
- */
-const filterOutLegacyDetectorsForNewerPeriods = (detectors, label = '', legacyDetectorNames = ['CPV', 'PHS']) => {
-    const shouldFilterOutLegacyDetectors = /LHC(2[5-9]|[3-9]\d)/.test(label ?? '');
-
-    return detectors.filter(({ name }) => !shouldFilterOutLegacyDetectors || !legacyDetectorNames.includes(name));
-};
-
 exports.ucFirst = ucFirst;
 
 exports.lcFirst = lcFirst;
@@ -99,5 +85,3 @@ exports.snakeToCamel = snakeToCamel;
 exports.snakeToPascal = snakeToPascal;
 
 exports.splitStringToStringsTrimmed = splitStringToStringsTrimmed;
-
-exports.filterOutLegacyDetectorsForNewerPeriods = filterOutLegacyDetectorsForNewerPeriods;
